fix: #938 user's ways: Markdown in Goal description shown incorrectly in Ways card view
Add hook useConvertingMarkdownToUrl to display the link in the user card, way card
