+++ mw-webapp/src/component/userCard/UserCard.tsx
@@ -9,6 +9,7 @@ import {PositionTooltip} from "src/component/tooltip/PositionTooltip";
 import {Tooltip} from "src/component/tooltip/Tooltip";
 import {VerticalContainer} from "src/component/verticalContainer/VerticalContainer";
 import {languageStore} from "src/globalStore/LanguageStore";
+import {useConvertingMarkdownToUrl} from "src/hooks/useConvertingMarkdownToUrl";
 import {
   UserNotSaturatedWay,
   UserTag,
@@ -39,6 +40,7 @@ interface UserCardProps {
  */
 export const UserCard = observer((props: UserCardProps) => {
   const {language} = languageStore;
+  const renderDescription = useConvertingMarkdownToUrl(props.userPreview.description);
 
   /**
    * Render way tags
@@ -54,49 +56,6 @@ export const UserCard = observer((props: UserCardProps) => {
           />
         ))}
       </HorizontalContainer>
-    );
-  };
-
-  /**
-   * The render Description function processes the description text to find markdown-style
-   * links and convert them into
-   * clickable anchor tags.
-   */
-  const renderDescription = (description: string) => {
-    const linkPattern = /\[([^\]]*)]\((https?:\/\/[^)]+)\)/g;
-    const urlGroupIndex = 2;
-    const urlGroupNumber = 1;
-
-    const parts = [];
-    let lastIndex = 0;
-    let match: RegExpExecArray | null;
-
-    while ((match = linkPattern.exec(description)) !== null) {
-      if (match.index > lastIndex) {
-        parts.push(description.substring(lastIndex, match.index));
-      }
-      let linkText = match[urlGroupNumber];
-      if (linkText.startsWith("**") && linkText.endsWith("**")) {
-        linkText = linkText.slice(urlGroupIndex, -urlGroupIndex);
-      }
-
-      parts.push(
-        <p key={match.index}>
-          {linkText || match[urlGroupIndex]}
-          {" "}
-        </p>,
-      );
-      lastIndex = linkPattern.lastIndex;
-    }
-
-    if (lastIndex < description.length) {
-      parts.push(description.substring(lastIndex));
-    }
-
-    return (
-      <p className={styles.description}>
-        {parts}
-      </p>
     );
   };
 
@@ -141,9 +100,11 @@ export const UserCard = observer((props: UserCardProps) => {
           </HorizontalContainer>
           <Tooltip
             position={PositionTooltip.BOTTOM}
-            content={renderDescription(props.userPreview.description)}
+            content={renderDescription}
           >
-            {renderDescription(props.userPreview.description)}
+            <p className={styles.description}>
+              {renderDescription}
+            </p>
           </Tooltip>
           {renderUserTags(props.userPreview.tags)}
         </VerticalContainer>
